backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Apoorv Yadav <apoorv.ydv17@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain-core = "^0.2.17"
typer = "^0.12.3"
langchain = "^0.2.7"
langchain-community = "^0.2.7"
python-dotenv = "^1.0.1"
pydantic = "^2.8.2"
uvicorn = "^0.30.1"
langgraph-cli = "^0.1.48"
fastapi = "^0.111.0"
langgraph = "^0.1.8"
langserve = "^0.2.2"
rich = "^13.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
